+++ CHANGES.md
@@ -1,3 +1,8 @@
 # Release Note
+### v1.1.0
+- Currently, the trigger function supports arguments.
+- No longer need to call manager.start() manually.
+- Fix multiple type hot key bug.
+___
 ### v1.0 - Initial Release
 - The first version.